Added (under testing) docs feature + refactored code (split into cogs)
