+++ config.py
@@ -1,0 +1,5 @@
+PREFIX = '!'
+TOKEN = ''
+#Todo -- update class dictionary 
+BLACKLIST_COGS = {}
+
